// microservices-kafka-elastisearch/order_service/src/dto/orderRequest.dto.ts
export type OrderLineItemType = {
	id: number;
	productId: number;
	itemName: string;
	qty: number;
	price: number;
	orderId: number;
	createdAt: Date;
	updatedAt: Date;
};

export interface OrderWithLineItems {
	id?: number;
	customerId: number;
	orderNumber: number;
	txnId: string | null;
	amount: string;
	status: string;
	orderItems: OrderLineItemType[];
	createdAt?: Date;
	updatedAt?: Date;
}